Update tests to expect `precision` property for table fields
N.B.: there are three pre-existing test failures. These failures exist in the master branch, before the changes in this branch were made. This commit does not attempt to fix those existing test failures, only the failures caused by the changes in this branch (in other words, it ensures things are no worse than they are in master).