Pass along field precision — FE-5228 (#61)
* In `getFieldsAsync()`, include column precision property

This is the precision of the column's data type, as reported by the backend via the Thrift bindings. This property is now included when returning the list of fields for a table, along with the pre-existing type, is_array, etc. properties.

* Update tests to expect `precision` property for table fields

N.B.: there are three pre-existing test failures. These failures exist in the master branch, before the changes in this branch were made. This commit does not attempt to fix those existing test failures, only the failures caused by the changes in this branch (in other words, it ensures things are no worse than they are in master).