Pre-compute current time in `replaceRelative()`
This is a fix for a failing unit test in Immerse. Because we're now working with millisecond-precision timestamps, our time tolerances are a lot tighter. Previously, the `replaceRelative()` function would compute the current time on-demand. However, in our unit tests, if we were comparing two `NOW()` values, these could be off my a millisecond or so. To fix this, `replaceRelative()` calculates the current time at the start of the function, and uses that throughout. This means `NOW()` should always be the same value.