[FE-4958] High Precision Timestamps (#85)
* Remove CAST from Date column filter and update TIMESTAMP filter to millisecond precision

* Add millisecond to TIME_LABEL hash

* Build dist file

* re-built

* Recompile

* Update timestamp logic to work at ms resolution

* Pre-compute current time in `replaceRelative()`

This is a fix for a failing unit test in Immerse. Because we're now working with millisecond-precision timestamps, our time tolerances are a lot tighter. Previously, the `replaceRelative()` function would compute the current time on-demand. However, in our unit tests, if we were comparing two `NOW()` values, these could be off my a millisecond or so. To fix this, `replaceRelative()` calculates the current time at the start of the function, and uses that throughout. This means `NOW()` should always be the same value.

* Let's try to solve this race condition again...

* Fix failing unit tests for high-precision timestamps

* Commit updated package-lock.json

* [FE-8853] Use higher precision when filtering date ranges (#87)

* Use higher precision when filtering date ranges

* test fixes

* Restore missing test due to bad merge